\name{ebd}
\alias{ebd}
\title{
Extrahepatic Cholangiocarcinoma Data
}
\description{
  The extrahepatic cholangiocarcinoma data comes form the US Surveilance, Epidemiology, and End Results (SEER) program of the National Cancer Institute. 
}
\references{
Hankey B., Ries L., and Edwards B. (1999). The surveillance, epidemiology, and end results program a national resource. \emph{Cencer Epidemiology Biomarkers and Prevention}, 12:1117-1121.
}
\usage{data(ebd)}
\keyword{datasets}

