\name{odcNews}
\alias{odcNews}
\title{Show the NEWS file of the OutlierDC package}
\description{
Show the NEWS file of the OutlierDC package.
}
\usage{
odcNews()
}
\seealso{
    \code{\link{OutlierDC-package}} \cr
    \code{\link{odc}}\cr 
    \code{\link{coef}}, \code{\link{plot}}, \code{\link{show}}, \code{\link{update}}
}
\examples{
odcNews()
}
\keyword{methods}
