\name{show}
\docType{methods}
\alias{show}
\alias{show,OutlierDC-method}
\title{ a show method for \code{OutlierDC}}
\description{
    This function provides a summary for the \code{OutlierDC} class. 
}
\usage{
    \S4method{show}{OutlierDC}(object)
}
\arguments{
    \item{object}{
        fitted object of class \code{\linkS4class{OutlierDC}}.  
    }
}
\details{
  This function is a method for the generic function \code{show} for the S4 class \code{OutlierDC}.  It can be invoked by calling print for an object of the appropriate class, or directly by calling \code{show} regardless of 
  the class of the object. 
}
\seealso{
    \code{\link{odc}} and \code{\linkS4class{OutlierDC}}
}
\keyword{methods}