\name{ebd}
\alias{ebd}
\docType{data}
\title{
	Extrahepatic Cholangiocarcinoma Data
}
\description{
  The extrahepatic cholangiocarcinoma data comes form the US Surveilance, Epidemiology, and End Results (SEER) program of the National Cancer Institute. 
}
\usage{data(ebd)}
\source{
	Hankey B., Ries L., and Edwards B. (1999). The surveillance, epidemiology, and end results program a national resource. \emph{Cencer Epidemiology Biomarkers and Prevention}, 12:1117-1121.
}
\examples{
	data(ebd)
}
\keyword{datasets}

